requirements.txt
cryptography==37.0.4
pwntools==4.8.0
clint==0.5.1